+++ console/src/dataSources/services/citus/index.tsx
@@ -68,6 +68,13 @@ export const supportedFeatures: DeepRequired<SupportedFeaturesType> = {
     nonTrackableFunctions: {
       enabled: true,
     },
+    modify: {
+      enabled: true,
+      comments: {
+        view: true,
+        edit: true,
+      },
+    },
   },
 };